BFF-1373 Fixed issue where the amount presets were not saved properly
