+++ lib/features/amount_presets/models/user_presets.dart
@@ -1,30 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:givt_app/features/amount_presets/models/preset.dart';
 
-class AmountPresets extends Equatable {
-  const AmountPresets({
-    required this.presets,
-  });
 
-  factory AmountPresets.fromJson(Map<String, dynamic> json) => AmountPresets(
-        presets: (json['amountPresets'] as List<dynamic>)
-            .map((e) => UserPresets.fromJson(e as Map<String, dynamic>))
-            .toList(),
-      );
-
-  const AmountPresets.empty() : presets = const [];
-
-  final List<UserPresets> presets;
-
-  Map<String, dynamic> toJson() => <String, dynamic>{
-        'amountPresets': presets.map((e) => e.toJson()).toList(),
-      };
-
-  @override
-  List<Object?> get props => [presets];
-
-  static const String tag = 'amountPresets';
-}
 
 class UserPresets extends Equatable {
   const UserPresets({
